Allow ICMP traffic to ourselves
Otherwise the ICMP port unreachables don't reach us.
These are sent to the public IP address of the interface, but come out of loopback
as the public IP is not known and might change, just allow every ICMP out of loopback.
ICMP traffic to other hosts is not sent out of loopback and not allowed by this rule.