Don't abort if a whitelist rule fails
If a whitelist rule failed (eg. because the DNS name referenced was not
resolveable) the script just stopped and did not add any further rules.
The new code only reports an error message but continues to load the
rest of the rules. If any rule failed, the return code is 1.

This also makes the script dependent on bash because POSIX shell
lacks some features needed.