Resolve hostnames before enableing the block policy
After the default policy is in place, DNS may no longer work because it
depends on rules not yet set. An example of this is DNS traffic which
goes through a tunnel which needs some rules to be in place.

The idea here is that the previous configuration hopefully allows these
DNS queries to succeed. If this is not the case there is nothing we can
do anyway.